[#111] feat : 로그인 토큰이 있을 시 myboard로 리다이렉트
