+++ package.json
@@ -15,6 +15,7 @@
     "next": "14.0.4",
     "next-auth": "^5.0.0-beta.4",
     "react": "^18.2.0",
+    "react-cookie": "^7.0.0",
     "react-datepicker": "^4.24.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.49.2",
