+++ src/app/(afterLogin)/dashboard/[dashboardId]/page.tsx
@@ -1,17 +1,17 @@
 'use client';
+import { useSetRecoilState, useRecoilState } from 'recoil';
+import { useEffect } from 'react';
+import { FieldValues } from 'react-hook-form';
+import { useRouter } from 'next/navigation';
+
 import { CardList } from '@/src/app/(afterLogin)/_component/CardList';
 import AddColumn from '@/src/app/_component/Button/AddColumn';
 import { axiosInstance } from '@/src/app/_util/axiosInstance';
-import { useEffect } from 'react';
-import { useRecoilState } from 'recoil';
-import { Column } from '../../_constant/type';
-import { useSetRecoilState } from 'recoil';
+import { Column } from '@/src/app/(afterLogin)/_constant/type';
 import { columnState, dashboardIdState } from '@/src/app/_recoil/cardAtom';
-import { FieldValues } from 'react-hook-form';
 import useRenderModal from '@/src/app/_hook/useRenderModal';
 import { MODALTYPE } from '@/src/app/_constant/modalType';
-import { getAccessToken } from '@/src/app/_util/axiosInstance';
-import { useRouter } from 'next/navigation';
+import { getAccessToken } from '@/src/app/_util/getAccessToken';
 
 export default function DashBoard({ params }: { params: { dashboardId: string } }) {
   const [columns, setColumns] = useRecoilState(columnState);
