+++ src/app/_util/getAccessToken.tsx
@@ -1,0 +1,15 @@
+import { UserDataType } from '../_constant/type';
+
+export const getAccessToken = () => {
+  let userDatas;
+  let accessToken = '';
+  if (typeof window !== 'undefined') {
+    userDatas = localStorage.getItem('taskifyUserData');
+  }
+
+  if (userDatas) {
+    const userData: UserDataType = JSON.parse(userDatas);
+    accessToken = userData.accessToken;
+    return accessToken;
+  }
+};
