--- src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,0 @@
-export { GET, POST } from '@/src/auth';
-
-//[...nextauth] => catch-all route
-
-//GET, POST, PATCH, DELETE 함수 등을 정의할 수 있다.
