--- src/middleware.txt
@@ -1,30 +1,0 @@
-// //auth.ts에 있는 auth 함수를 middleware로 선언해줌으로써 export default middleware(auth)를 해준 것이나 다름 없다.
-
-// // export { auth as middleware } from '@/src/auth';
-
-// import { NextResponse } from 'next/server';
-// import type { NextRequest } from 'next/server';
-
-// //해당 페이지로 접근할 때 middleware 함수를 먼저 실행한다.
-
-// export function middleware(request: NextRequest) {
-//   // 클라이언트 사이드에서만 실행되는 코드
-//   if (typeof window !== 'undefined') {
-//     // 로컬 스토리지에서 accessToken 가져오기
-//     const accessToken = localStorage.getItem('accessToken');
-
-//     // accessToken이 없으면 로그인 페이지로 리다이렉트
-//     if (!accessToken) {
-//       // 로그인 페이지의 경로를 설정하세요.
-//       const loginPageUrl = '/login';
-//       return NextResponse.redirect(loginPageUrl);
-//     }
-//   }
-
-//   // 서버 사이드에서는 다음으로 진행
-//   return NextResponse.next();
-// }
-
-// export const config = {
-//   matcher: ['/mypage', '/myboard', '/dashboard/:id*'],
-// };
